continuacion de nueva arquitectura

+++ server/functions/delete/usuarios.php
@@ -4,9 +4,29 @@ function eliminar_admin()
 {
     include_once '../conexion.php';
 
-    $id_admin = $_POST['id_admin'];
+    $respuesta = 
+    [
+      'titulo'=> 'Ups',
+      'cuerpo'=> 'No Pudimos Procesar Su Solicitud',
+      'accion'=> 'warning'
+    ];
 
-    $deletesql = 'DELETE FROM usuarios WHERE Id=?';
-    $sentenceDelete = $pdo->prepare($deletesql);
-    $sentenceDelete-> execute(array($id_admin));
+    if(isset($_POST['id_usuario']))
+    {
+        $id_usuario = $_POST['id_usuario'];
+
+        $deletesql = 'DELETE FROM usuarios WHERE Id=?';
+        $sentenceDelete = $pdo->prepare($deletesql);
+        if($sentenceDelete-> execute(array($id_usuario)))
+        {
+            $respuesta = 
+            [
+              'titulo'=> 'Operación Exitosa',
+              'cuerpo'=> '',
+              'accion'=> 'success'
+            ];
+        }
+
+        echo json_encode($respuesta);
+    }
 }
